fix(npm): Fix for empty NPM tarballs generated by semantic-release.
This is my second time running into to this; should have known better.
Still — I'm [not the first][github] to have found this a little
unintuitive -- it [appears][github 2] that not even `prepack` /
`prepublishOnly` scripts get run during the `@semantic-release/npm`
`publish` step (which is consistent with what I've observed in CI).

`¯\_(ツ)_/¯`

[github]: https://github.com/semantic-release/npm/issues/97
[github 2]: https://github.com/semantic-release/npm/issues/114#issuecomment-437451329